pyproject.toml
[tool.poetry]
authors = ["anonymous"]
description = ""
name = "netevolve"
version = "0.1.0"
[tool.mypy]
# エラー時のメッセージを詳細表示
show_error_context = true
# エラー発生箇所の行数/列数を表示
show_column_numbers = true
# import 先のチェックを行わない (デフォルトだとサードパーティーライブラリまでチェックする)
ignore_missing_imports = true
# 関数定義の引数/戻り値に型アノテーション必須
disallow_untyped_defs = true
# デフォルト引数に None を取る場合型アノテーションに Optional 必須
no_implicit_optional = true
# 戻り値が Any 型ではない関数の戻り値の型アノテーションが Any のとき警告
warn_return_any = true
# mypy エラーに該当しない箇所に `# type: ignore` コメントが付与されていたら警告
# ※ `# type: ignore` が付与されている箇所は mypy のエラーを無視出来る
warn_unused_ignores = true
# 冗長なキャストに警告
warn_redundant_casts = true

[tool.black]
line-length = 79

[tool.isort]
import_heading_firstparty = "First Party Library"
import_heading_localfolder = "Local Library"
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third Party Library"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"

[tool.flake8]
exclude = [".venv", ".git", "__pycache__"]
extend-ignore = ["E203", "W503"]
max-complexity = 10
max-line-length = 79
[tool.poetry.dependencies]
cloudpickle = "^2.2.0"
dill = "^0.3.6"
gensim = "^4.3.0"
matplotlib = "^3.5.2"
networkx = "^2.8.2"
numpy = "^1.22.4"
ogb = "^1.3.5"
pandas = "^1.5.2"
python = ">=3.9,<3.12"
scikit-learn = "^1.1.3"
scipy = "^1.8.1"
torchviz = "^0.0.2"
tqdm = "^4.64.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"
flake8-builtins = "^2.1.0"
flake8-isort = "^6.0.0"
flake8-pytest-style = "^1.7.2"
flake8-unused-arguments = "^0.0.13"
isort = "^5.12.0"
mypy = "^1.0.1"
pep8-naming = "^0.13.3"
pyproject-flake8 = "^6.0.0.post1"
tdqm = "^0.0.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
